Replace Article by WikiPage, and related changes.
Article has been deprecated for a while and was dropped from mw 1.29
altogether. The getContent() for WikiPage returns the Content
instance instead of plain string, unlike getContent() for Article. As
a consequence, we need to use methods for Content to get the
appropriate data. This would typically be done using Content's
getParserOutput() to get the Parser data, and then returning the
output from running getText() on that. This works perfectly for the
Footer, but it doesn't for the SubNav bar, which expects a plain text
list. To get this, we simply return the getNativeData() from the
Content, without running the parser on it. This seems to work for
IFPA, but seems less than ideal, and would likely run into problems
with more complicated situations such as when dynamic menus are used.